requirements.txt
python_telegram_bot==21.2
spotipy==2.23.0
beautifulsoup4==4.12.3
click==8.1.3

